config: Yale YDM3109A Smart Lock device support
Basic device support works without this device file, but this adds
proper device identification and translation of a collection of alerts,
mostly using the Yale templates:

* alarm_map_keypad_limit - works
* alarm_map_manual_unlock - works
* alarm_map_rf_unlock - works
* alarm_map_keypad_unlock - works, including user slot used
* alarm_map_manual_lock - works
* alarm_map_rf_lock - works
* alarm_map_auto_relock - works
* alarm_map_rfid_unlock - works, including user slot used, but due to no
  specific Z-wave notification type for "opened by RFID fob" this comes
  through as a normal keypad unlock

Notifications that are sent by the device and translated to status
values:

* alarm_map_deadbolt_jammed
* alarm_map_low_battery

Could not figure out a way to trigger these, but have left in the file
anyway:

* alarm_map_keypad_lock
* alarm_map_critical_battery

Setting/clearing user PIN codes works using zwave_js.set_lock_usercode /
zwave_js.clear_lock_usercode in Home Assistant.

These devices have an interior "Forced Lock/Privacy Switch" that
disables the external keypad when in the "ON" position. A notification
with the current state is sent when the device is powered on, and again
whenver the switch position is changed.

Alarm mappings have been added to translate this to a "Home Monitoring"
/ "Home occupied" notification, as there is nothing suitable under the
Access Control (0x06) section of the Z-Wave notification registry as of
the 17/08/2023 release.