feat: add rate of incoming messages/second for nodes
Keeps track of the rate of incoming messages for each node, averaged over the last minute.

Implemented using a new `RateTracker` class. Uses `setTimeout` to recalculate if no messages are received for a while. To prevent `Timeout` overload, the rate is only recalculated every 2 seconds, even if messages are received more frequently.